+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `IterableLens#mapping`, an `Iso` that maps values
+
 ### Changed
 - `TypeSafeKey.Simple` now has a default `#apply` implementation